#17 Move directory and change ToolWindow ID
