+++ src/main/resources/META-INF/plugin.xml
@@ -18,10 +18,10 @@
 
     <extensions defaultExtensionNs="com.intellij">
         <toolWindow
-                factoryClass="org.jqassistant.tooling.intellij.plugin.editor.effectiveConfig.EffectiveConfigToolWindowFactory"
-                id="EffectiveConfigToolWindow"
-                anchor="bottom"
-                icon="/icons/jqaShortIcon_mono.svg"/>
+            factoryClass="org.jqassistant.tooling.intellij.plugin.editor.config.toolwindow.EffectiveConfigToolWindowFactory"
+            id="Effective Configuration"
+            anchor="bottom"
+            icon="/icons/jqaShortIcon_mono.svg"/>
         <toolWindow id="jQAssistant Report" secondary="true" icon="AllIcons.Actions.ProjectWideAnalysisOff"
                     anchor="right"
                     factoryClass="org.jqassistant.tooling.intellij.plugin.editor.report.ReportToolWindowFactory"/>
